Fix memory consumption, Restructure for Node.js and Gulp.js
