+++ .gitignore
@@ -1,1 +1,2 @@
 localtesting/*
+node_modules/
