+++ package.json
@@ -1,0 +1,28 @@
+{
+	"name": "codemirror-spell-checker",
+	"version": "1.0.5",
+	"description": "Dead-simple spell checking in CodeMirror.",
+	"keywords": ["codemirror", "code", "mirror", "spell", "checker", "checking", "spellchecker", "javascript"],
+	"homepage": "https://github.com/NextStepWebs/codemirror-spell-checker",
+	"main": "gulpfile.js",
+	"license": "MIT",
+	"company": "Next Step Webs, Inc.",
+	"author": {
+		"name": "Wes Cossick",
+		"url": "http://www.WesCossick.com"
+	},
+	"bugs": {
+		"url": "https://github.com/NextStepWebs/codemirror-spell-checker/issues"
+	},
+	"dependencies": {
+		"gulp": "*",
+		"gulp-minify-css": "*",
+		"gulp-uglify": "*",
+		"gulp-concat": "*",
+		"gulp-header": "*"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/NextStepWebs/codemirror-spell-checker"
+	}
+}
