--- source files/README.md
@@ -1,9 +1,0 @@
-# How to compile
-Minify the JS in this order:
-
-1. `typo/typo.js`
-1. `spell-checker.js`
-
-Minify the CSS in this order:
-
-1. `spell-checker.css`
